pyclima-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
numpy>=1.26.0
openai>=1.0.0
python-dotenv>=1.0.0
xarray
netCDF4
h5netcdf
python-multipart
scipy
pandas
